requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.10.0
    # via django
charset-normalizer==3.4.3
    # via reportlab
cssselect2==0.8.0
    # via svglib
django==5.2.7
    # via
    #   -r requirements.in
    #   django-classy-tags
    #   django-cms
    #   django-entangled
    #   django-filer
    #   django-formtools
    #   django-parler
    #   django-polymorphic
    #   django-sekizai
    #   django-treebeard
    #   djangocms-admin-style
    #   djangocms-versioning
    #   easy-thumbnails
django-classy-tags==4.1.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==5.0.4
    # via
    #   -r requirements.in
    #   djangocms-alias
    #   djangocms-attributes-field
    #   djangocms-frontend
    #   djangocms-link
    #   djangocms-simple-admin-style
    #   djangocms-text-ckeditor
    #   djangocms-versioning
django-entangled==0.6.3
    # via djangocms-frontend
django-environ==0.12.0
    # via -r requirements.in
django-filer==3.3.2
    # via
    #   -r requirements.in
    #   djangocms-frontend
django-formtools==2.5.1
    # via django-cms
django-fsm==2.8.2
    # via
    #   -r requirements.in
    #   djangocms-versioning
django-parler==2.3
    # via
    #   -r requirements.in
    #   djangocms-alias
django-polymorphic==4.1.0
    # via django-filer
django-sekizai==4.1.0
    # via django-cms
django-treebeard==4.7.1
    # via django-cms
djangocms-admin-style==3.3.1
    # via django-cms
djangocms-alias==3.0.1
    # via -r requirements.in
djangocms-attributes-field==4.1.1
    # via
    #   djangocms-frontend
    #   djangocms-link
djangocms-frontend==2.2.0
    # via -r requirements.in
djangocms-link==5.1.0
    # via djangocms-frontend
djangocms-simple-admin-style==1.1.5
    # via -r requirements.in
djangocms-text-ckeditor==5.1.7
    # via -r requirements.in
djangocms-versioning==2.4.0
    # via -r requirements.in
easy-thumbnails[svg]==2.10.1
    # via
    #   django-filer
    #   djangocms-frontend
freetype-py==2.5.1
    # via rlpycairo
html5lib==1.1
    # via djangocms-text-ckeditor
lxml==6.0.2
    # via svglib
packaging==25.0
    # via
    #   django-cms
    #   djangocms-text-ckeditor
    #   djangocms-versioning
pillow==11.3.0
    # via
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
pycairo==1.28.0
    # via rlpycairo
reportlab==4.4.4
    # via
    #   easy-thumbnails
    #   svglib
rlpycairo==0.4.0
    # via svglib
six==1.17.0
    # via html5lib
sqlparse==0.5.3
    # via django
svglib==1.6.0
    # via easy-thumbnails
tinycss2==1.4.0
    # via
    #   cssselect2
    #   svglib
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2

# The following packages are considered to be unsafe in a requirements file:
# setuptools
